chemical_visualizer/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
pandas==2.1.3
reportlab==4.0.7
celery==5.3.4
redis==5.0.1
channels==4.0.0
channels-redis==4.1.0
gunicorn==21.2.0
daphne==4.0.0
# AI Enhancement Dependencies
google-generativeai==0.3.2
scipy==1.11.4
python-dotenv==1.0.0
# PDF Enhancement Dependencies
matplotlib==3.8.2
seaborn==0.13.0
# Cache Dependencies
django-redis==5.4.0


